Separate user models more cleanly to fix tests
