+++ app/Models/NativeUser.php
@@ -2,6 +2,8 @@
 
 namespace App\Models;
 
+use App\Notifications\ResetPasswordNotification;
+use App\Notifications\VerifyEmailNotification;
 use Illuminate\Auth\MustVerifyEmail;
 use Illuminate\Auth\Passwords\CanResetPassword;
 use Illuminate\Contracts\Auth\CanResetPassword as CanResetPasswordContract;
@@ -10,14 +12,12 @@ use Tightenco\Parental\HasParent;
 
 /**
  * @property int $id
- * @property int $course_id
  * @property string $name
  * @property string $group
  * @property string $password
  * @property string $email
- * @property string $salt
  * @property string $image_url
- * @property string $type
+ * @property string $login_provider
  * @property Observation[] $observations
  * @property Course[] $courses
  * @property Course[] $nonArchivedCourses
@@ -29,4 +29,46 @@ use Tightenco\Parental\HasParent;
 class NativeUser extends User implements CanResetPasswordContract, MustVerifyEmailContract
 {
     use HasParent, CanResetPassword, MustVerifyEmail;
+
+    public function __construct(...$args) {
+        parent::__construct(...$args);
+
+        $this->casts['email_verified_at'] = 'datetime';
+    }
+
+    /**
+     * @return \Illuminate\Database\Eloquent\Relations\HasMany
+     */
+    public function loginAttempts()
+    {
+        return $this->hasMany('App\Models\LoginAttempt');
+    }
+
+    /**
+     * @return \Illuminate\Database\Eloquent\Relations\HasMany
+     */
+    public function recoveryAttempts()
+    {
+        return $this->hasMany('App\Models\RecoveryAttempt');
+    }
+
+    /**
+     * Send the password reset notification.
+     *
+     * @param  string  $token
+     * @return void
+     */
+    public function sendPasswordResetNotification($token)
+    {
+        $this->notify(new ResetPasswordNotification($token));
+    }
+
+    /**
+     * Send the email verification notification.
+     *
+     * @return void
+     */
+    public function sendEmailVerificationNotification() {
+        $this->notify(new VerifyEmailNotification());
+    }
 }
