+++ app/Models/User.php
@@ -2,8 +2,6 @@
 
 namespace App\Models;
 
-use App\Notifications\ResetPasswordNotification;
-use App\Notifications\VerifyEmailNotification;
 use Carbon\CarbonInterface;
 use Illuminate\Auth\Authenticatable;
 use Illuminate\Auth\MustVerifyEmail;
@@ -19,12 +17,10 @@ use Tightenco\Parental\HasChildren;
 
 /**
  * @property int $id
- * @property int $course_id
  * @property string $name
  * @property string $group
  * @property string $password
  * @property string $email
- * @property string $salt
  * @property string $image_url
  * @property string $login_provider
  * @property Observation[] $observations
@@ -32,8 +28,6 @@ use Tightenco\Parental\HasChildren;
  * @property Course[] $nonArchivedCourses
  * @property Course[] $archivedCourses
  * @property Course $last_accessed_course
- * @property LoginAttempt[] $loginAttempts
- * @property RecoveryAttempt[] $recoveryAttempts
  */
 class User extends Model implements AuthenticatableContract, AuthorizableContract, CanResetPasswordContract, MustVerifyEmailContract
 {
@@ -50,13 +44,6 @@ class User extends Model implements AuthenticatableContract, AuthorizableContrac
      * @var array
      */
     protected $hidden = ['password', 'remember_token'];
-
-    /**
-     * The attributes that should be cast to native types.
-     *
-     * @var array
-     */
-    protected $casts = ['email_verified_at' => 'datetime'];
 
     /**
      * Name of the database column holding the login provider name
@@ -133,41 +120,5 @@ class User extends Model implements AuthenticatableContract, AuthorizableContrac
      */
     public function setLastUsedBlockDate($value, Course $course) {
         $this->courses()->updateExistingPivot($course->id, ['last_used_block_date' => Carbon::parse($value)]);
-    }
-
-    /**
-     * @return \Illuminate\Database\Eloquent\Relations\HasMany
-     */
-    public function loginAttempts()
-    {
-        return $this->hasMany('App\Models\LoginAttempt');
-    }
-
-    /**
-     * @return \Illuminate\Database\Eloquent\Relations\HasMany
-     */
-    public function recoveryAttempts()
-    {
-        return $this->hasMany('App\Models\RecoveryAttempt');
-    }
-
-    /**
-     * Send the password reset notification.
-     *
-     * @param  string  $token
-     * @return void
-     */
-    public function sendPasswordResetNotification($token)
-    {
-        $this->notify(new ResetPasswordNotification($token));
-    }
-
-    /**
-     * Send the email verification notification.
-     *
-     * @return void
-     */
-    public function sendEmailVerificationNotification() {
-        $this->notify(new VerifyEmailNotification());
     }
 }
