+++ database/factories/NativeUserFactory.php
@@ -1,6 +1,6 @@
 <?php
 
-use App\Models\User;
+use App\Models\NativeUser;
 use Faker\Generator as Faker;
 use Illuminate\Support\Str;
 
@@ -15,7 +15,7 @@ use Illuminate\Support\Str;
 |
 */
 
-$factory->define(User::class, function (Faker $faker) {
+$factory->define(NativeUser::class, function (Faker $faker) {
     return [
         'name' => $faker->firstName,
         'email' => $faker->unique()->safeEmail,
